Stores user's board when Recovery mode ends
Added a method to the blockrain.src that returns the board as a string
since that is what the method to create a board uses. It had a method
that returned the board as an array but that isn't helpful. Also added
posting your game's result to '/recovery/board' when the game ends and
the server stores that board as a variable and sends it when a get
request is sent to '/recovery/board'. We can move storing it to a DB
later.